Wired through deleting a node with cascade delete of edges
