+++ src/lib/actions/sandbox-actions.ts
@@ -2,6 +2,9 @@
 import { Position } from '@xyflow/react';
 import { type Node } from '@xyflow/react';
 import { type Edge } from '@xyflow/react';
+import { Config } from '@config/constants';
+
+const { WorkflowNode } = Config;
 
 // these are hardcoded nodes to be replaced with nodes from the db
 const position = { x: 50, y: 100 };
@@ -9,7 +12,7 @@ const automationNodes = [
     {
         id: 'horizontal-1',
         sourcePosition: Position.Right,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 1' },
         position
     },
@@ -17,7 +20,7 @@ const automationNodes = [
         id: 'horizontal-2',
         sourcePosition: Position.Right,
         targetPosition: Position.Left,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 2' },
         position
     },
@@ -25,7 +28,7 @@ const automationNodes = [
         id: 'horizontal-3',
         sourcePosition: Position.Right,
         targetPosition: Position.Left,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 3' },
         position
     },
@@ -33,7 +36,7 @@ const automationNodes = [
         id: 'horizontal-4',
         sourcePosition: Position.Right,
         targetPosition: Position.Left,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 4' },
         position
     },
@@ -41,7 +44,7 @@ const automationNodes = [
         id: 'horizontal-5',
         sourcePosition: Position.Top,
         targetPosition: Position.Bottom,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 5' },
         position
     },
@@ -49,7 +52,7 @@ const automationNodes = [
         id: 'horizontal-6',
         sourcePosition: Position.Bottom,
         targetPosition: Position.Top,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 6' },
         position
     },
@@ -57,7 +60,7 @@ const automationNodes = [
         id: 'horizontal-7',
         sourcePosition: Position.Right,
         targetPosition: Position.Left,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 7' },
         position
     },
@@ -65,7 +68,7 @@ const automationNodes = [
         id: 'horizontal-8',
         sourcePosition: Position.Right,
         targetPosition: Position.Left,
-        type: 'automationNode',
+        type: WorkflowNode.AutomationNode,
         data: { label: 'Node 8' },
         position
     }
@@ -77,49 +80,49 @@ const automationEdges = [
         id: 'horizontal-e1-2',
         source: 'horizontal-1',
         target: 'horizontal-2',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e1-3',
         source: 'horizontal-1',
         target: 'horizontal-3',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e2-4',
         source: 'horizontal-2',
         target: 'horizontal-4',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e3-5',
         source: 'horizontal-3',
         target: 'horizontal-5',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e3-6',
         source: 'horizontal-3',
         target: 'horizontal-6',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e5-7',
         source: 'horizontal-5',
         target: 'horizontal-7',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: true
     },
     {
         id: 'horizontal-e6-8',
         source: 'horizontal-6',
         target: 'horizontal-8',
-        type: 'automationEdge',
+        type: WorkflowNode.AutomationEdge,
         animated: false
     }
 ] as Edge[];
